+++ CHANGELOG.md
@@ -8,7 +8,7 @@
 
 - [Send] Remove transaction confirmation check [#2072](https://github.com/thorchain/asgardex-electron/issues/2072)
 - [Ledger] Disable selecting of slippage tolerance for Ledger/BTC temporary [#2068](https://github.com/thorchain/asgardex-electron/issues/2068)
-- [BTC] Support confirmed balances only #2066[](https://github.com/thorchain/asgardex-electron/pull/2066)
+- [BTC] Support confirmed balances only [#2066](https://github.com/thorchain/asgardex-electron/pull/2066)
 - Shorten memos [#2052](https://github.com/thorchain/asgardex-electron/issues/2052)
 - Re-enable testnet for releases (production) [#2048](https://github.com/thorchain/asgardex-electron/issues/2048)
 - Update for RU translation (better translation and fix for typos) by @TreefeedXavier [#2046](https://github.com/thorchain/asgardex-electron/pull/2046)
@@ -16,7 +16,8 @@
 
 ## Fix
 
-- [i18n] Translate pooled #2047
+- [i18n] Translate pooled [#2047](https://github.com/thorchain/asgardex-electron/issues/2047)
+- [BTC] Broadcasting a tx to Haskoin might end in 500 error [#2077](https://github.com/thorchain/asgardex-electron/issues/2077)
 
 # 0.7.1 (2022-01-21)