+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 ## Update
 
+- [Send] Fix out of gas for RUNE txs [2081](https://github.com/thorchain/asgardex-electron/issues/2081)
+- Handle (confirmed) BTC balances the better way [#2082](https://github.com/thorchain/asgardex-electron/pull/2082)
 - [Send] Remove transaction confirmation check [#2072](https://github.com/thorchain/asgardex-electron/issues/2072)
 - [Ledger] Disable selecting of slippage tolerance for Ledger/BTC temporary [#2068](https://github.com/thorchain/asgardex-electron/issues/2068)
 - [BTC] Support confirmed balances only [#2066](https://github.com/thorchain/asgardex-electron/pull/2066)